refactor: Localización mappers de economic data
