+++ src/main/java/com/helloworld/renting/mapper/economicalData/EconomicDataSelfEmployedMapper.java
@@ -1,4 +1,4 @@
-package com.helloworld.renting.mapper;
+package com.helloworld.renting.mapper.economicalData;
 
 import com.helloworld.renting.entities.EconomicDataSelfEmployed;
 import org.apache.ibatis.annotations.*;
