+++ src/test/java/com/helloworld/renting/service/economicdata/EconomicDataServiceTest.java
@@ -5,8 +5,8 @@ import com.helloworld.renting.entities.EconomicDataSelfEmployed;
 import com.helloworld.renting.exceptions.RentingException;
 import com.helloworld.renting.exceptions.db.DBException;
 import com.helloworld.renting.exceptions.notfound.EconomicDataNotFoundException;
-import com.helloworld.renting.mapper.EconomicDataEmployedMapper;
-import com.helloworld.renting.mapper.EconomicDataSelfEmployedMapper;
+import com.helloworld.renting.mapper.economicalData.EconomicDataEmployedMapper;
+import com.helloworld.renting.mapper.economicalData.EconomicDataSelfEmployedMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
